Remove dead code and reorganize slightly
