+++ src/scripts/modules/highlight.ts
